[feat] 시그널링 서버 상태 적용
